+++ summon-core/src/jvmMain/kotlin/code/yousef/summon/integration/springboot/WebFluxRenderer.kt
@@ -1,7 +1,6 @@
 package code.yousef.summon.integration.springboot
 
 import code.yousef.summon.annotation.Composable
-import code.yousef.summon.runtime.CallbackRegistry
 import code.yousef.summon.runtime.PlatformRenderer
 import code.yousef.summon.runtime.clearPlatformRenderer
 import code.yousef.summon.runtime.setPlatformRenderer
@@ -55,7 +54,6 @@ class WebFluxRenderer {
                     }
                 }
             } finally {
-                CallbackRegistry.clear()
                 clearPlatformRenderer()
             }
         }.subscribeOn(Schedulers.boundedElastic())
@@ -84,7 +82,6 @@ class WebFluxRenderer {
                     }
                 }
             } finally {
-                CallbackRegistry.clear()
                 clearPlatformRenderer()
             }
         }.subscribeOn(Schedulers.boundedElastic())